requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
httpx==0.27.0
python-multipart==0.0.9